requirements.txt
Django>=3.1.7
Pillow>=8.1.2
numpy>=1.20.1
pip>=21.0.1
parso>=0.8.1
jedi>=0.18.0
pytz>=2021.1
setuptools>=54.2.0
sqlparse>=0.4.1
asgiref>=3.3.1
Pygments>=2.8.1
ipython>=7.22.0
traitlets>=5.0.5
decorator>=4.4.2
backcall>=0.2.0
pickleshare>=0.7.5
pexpect>=4.8.0
wcwidth>=0.2.5
ptyprocess>=0.7.0
gunicorn>=20.0.4
psycopg2-binary
requests~=2.25.1

django-environ